requirements.txt
Flask>=3.0,<4.0
argon2-cffi==23.1.0